app/requirements.txt
fastapi
uvicorn 
requests
./boto3-1.28.21-py3-none-any.whl
./botocore-1.31.21-py3-none-any.whl
python-dotenv